requirements.txt
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0
chromadb==0.4.18
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.4
numpy==1.24.3
lxml==4.9.3
nltk==3.8.1
